changed channel.js  channelmessagesview.js
